+++ Cargo.toml
@@ -27,6 +27,7 @@ minreq = { version = "=2.2.1", features = ["https", "json-using-serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 training_mod_consts = { path = "training_mod_consts" }
+training_mod_tui = { path = "training_mod_tui" }
 
 [patch.crates-io]
 ring = { git = "https://github.com/skyline-rs/ring", branch = "0.16.20" }
