+++ training_mod_consts/Cargo.toml
@@ -10,4 +10,11 @@ strum_macros = "0.21.0"
 num = "0.4.0"
 num-derive = "0.3"
 num-traits = "0.2"
-skyline_smash = { git = "https://github.com/ultimate-research/skyline-smash.git" } 
+ramhorns = "0.12.0"
+paste = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+skyline_smash = { git = "https://github.com/ultimate-research/skyline-smash.git", optional = true }
+
+[features]
+default = ["smash"]
+smash = ["skyline_smash"]
