+++ training_mod_tui/src/lib.rs
@@ -1,0 +1,151 @@
+// use crossterm::{
+//     event::{self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode},
+//     execute,
+//     terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
+// };
+use std::{
+    error::Error,
+    io,
+    time::{Duration, Instant},
+};
+use std::borrow::Borrow;
+use tui::{
+    backend::{Backend},
+    layout::{Constraint, Corner, Direction, Layout},
+    style::{Color, Modifier, Style},
+    text::{Span, Spans},
+    widgets::{Block, Borders, List, ListItem, ListState},
+    Frame,
+};
+use tui::buffer::Buffer;
+
+pub use tui::{backend::TestBackend, Terminal};
+
+pub struct StatefulList<T> {
+    state: ListState,
+    items: Vec<T>,
+}
+
+impl<T> StatefulList<T> {
+    pub fn with_items(items: Vec<T>) -> StatefulList<T> {
+        StatefulList {
+            state: ListState::default(),
+            items,
+        }
+    }
+
+    pub fn next(&mut self) {
+        let i = match self.state.selected() {
+            Some(i) => {
+                if i >= self.items.len() - 1 {
+                    0
+                } else {
+                    i + 1
+                }
+            }
+            None => 0,
+        };
+        self.state.select(Some(i));
+    }
+
+    pub fn previous(&mut self) {
+        let i = match self.state.selected() {
+            Some(i) => {
+                if i == 0 {
+                    self.items.len() - 1
+                } else {
+                    i - 1
+                }
+            }
+            None => 0,
+        };
+        self.state.select(Some(i));
+    }
+
+    pub fn unselect(&mut self) {
+        self.state.select(None);
+    }
+}
+
+/// This struct holds the current state of the app. In particular, it has the `items` field which is a wrapper
+/// around `ListState`. Keeping track of the items state let us render the associated widget with its state
+/// and have access to features such as natural scrolling.
+///
+/// Check the event handling at the bottom to see how to change the state on incoming events.
+/// Check the drawing logic for items on how to specify the highlighting style for selected items.
+pub struct App<'a> {
+    pub items: StatefulList<(&'a str, &'a str)>,
+    pub events: Vec<(&'a str, &'a str)>,
+}
+
+impl<'a> App<'a> {
+    pub fn new() -> App<'a> {
+        App {
+            items: StatefulList::with_items(vec![
+                ("Mash Toggles", "Blargh"),
+                ("DI Direction", "Blargh"),
+            ]),
+            events: vec![
+                ("Event1", "INFO"),
+            ],
+        }
+    }
+
+    /// Rotate through the event list.
+    /// This only exists to simulate some kind of "progress"
+    pub fn on_tick(&mut self) {
+        let event = self.events.remove(0);
+        self.events.push(event);
+    }
+}
+
+pub fn ui<B: Backend>(f: &mut Frame<B>, app: &mut App) {
+    // Create two chunks with equal horizontal screen space
+    let chunks = Layout::default()
+        .direction(Direction::Horizontal)
+        .constraints([Constraint::Percentage(50), Constraint::Percentage(50)].as_ref())
+        .split(f.size());
+
+    // Iterate through all elements in the `items` app and append some debug text to it.
+    let items: Vec<ListItem> = app
+        .items
+        .items
+        .iter()
+        .map(|i| {
+            let mut lines = vec![Spans::from(i.0)];
+            ListItem::new(lines).style(Style::default().fg(Color::Black).bg(Color::White))
+        })
+        .collect();
+
+    // Create a List from all list items and highlight the currently selected one
+    let items = List::new(items)
+        .block(Block::default().title("List"))
+        .highlight_style(
+            Style::default()
+                .bg(Color::LightGreen)
+                .add_modifier(Modifier::BOLD),
+        )
+        .highlight_symbol(">> ");
+
+    let selected = app
+        .items
+        .state
+        .selected()
+        .map(|selected_index| app.items.items.get(selected_index).unwrap().1)
+        .unwrap_or("");
+
+    // We can now render the item list
+    f.render_stateful_widget(items, chunks[0], &mut app.items.state);
+
+    let values_list = List::new(vec![
+        ListItem::new(vec![
+                Spans::from("-".repeat(chunks[1].width as usize)),
+                Spans::from("Value"),
+                Spans::from(""),
+                Spans::from(selected),
+            ])
+        ])
+        .block(Block::default().title("List"))
+        .start_corner(Corner::TopLeft);
+    f.render_widget(values_list, chunks[1]);
+}
